requirements.txt
requests
python-dotenv
openai>=1.2.0
